Silence -Wformat warnings in lvgl_camera_display.cpp
On the ESP32-P4 (RISC-V ILP32) uint32_t is 'long unsigned int' and
lv_coord_t is 'long int', so the "%u"/"%d" specifiers used for
update_interval_, the frame-interval target, get_buffer_index(), the
draw-buf stride/size and the canvas width/height triggered -Wformat.
Cast each argument to (unsigned)/(int) to match, without changing any log
text or values.

Verified: esphome compile on 2026.7.0-dev builds green with 0 -Wformat
warnings.

+++ components/lvgl_camera_display/lvgl_camera_display.cpp
@@ -41,7 +41,7 @@ void LVGLCameraDisplay::setup() {
   ESP_LOGI(TAG, "LVGL Camera Display initialise (not started yet)");
   ESP_LOGI(TAG, "   Camera: Operationnelle");
   ESP_LOGI(TAG, "   Update interval: %u ms (~%d FPS) via LVGL timer",
-           this->update_interval_, 1000 / this->update_interval_);
+           (unsigned) this->update_interval_, (int) (1000 / this->update_interval_));
   ESP_LOGI(TAG, "Turn on the 'LVGL Camera Display' switch to start");
 }
 
@@ -151,7 +151,7 @@ void LVGLCameraDisplay::update_camera_frame_() {
 
       ESP_LOGI(TAG, "=== BENCHMARK (100 frames) ===");
       ESP_LOGI(TAG, "  FPS: %.1f | CPU: %.1f%%", fps, cpu_percent);
-      ESP_LOGI(TAG, "  Frame interval: %.1fms (target: %ums)", avg_frame_interval, this->update_interval_);
+      ESP_LOGI(TAG, "  Frame interval: %.1fms (target: %ums)", avg_frame_interval, (unsigned) this->update_interval_);
       ESP_LOGI(TAG, "  CPU time: %.1fms (capture: %.1fms + canvas: %.1fms)", avg_cpu_time, avg_capture, avg_canvas);
       ESP_LOGI(TAG, "  LVGL overhead: %.1fms | Skip: %.1f%%", lvgl_overhead, skip_rate);
 
@@ -197,8 +197,8 @@ void LVGLCameraDisplay::set_stats_label(lv_obj_t *label) {
 
 void LVGLCameraDisplay::dump_config() {
   ESP_LOGCONFIG(TAG, "LVGL Camera Display:");
-  ESP_LOGCONFIG(TAG, "  Update interval: %u ms", this->update_interval_);
-  ESP_LOGCONFIG(TAG, "  FPS cible: ~%d", 1000 / this->update_interval_);
+  ESP_LOGCONFIG(TAG, "  Update interval: %u ms", (unsigned) this->update_interval_);
+  ESP_LOGCONFIG(TAG, "  FPS cible: ~%d", (int) (1000 / this->update_interval_));
   ESP_LOGCONFIG(TAG, "  Canvas configure: %s", this->canvas_obj_ ? "OUI" : "NON");
 }
 
@@ -270,7 +270,7 @@ void LVGLCameraDisplay::update_canvas_() {
     this->is_canvas_ = lv_obj_check_type(this->canvas_obj_, &lv_canvas_class);
     ESP_LOGI(TAG, "Premier update - Widget type: %s", this->is_canvas_ ? "CANVAS" : "IMAGE");
     ESP_LOGI(TAG, "   Dimensions: %ux%u", width, height);
-    ESP_LOGI(TAG, "   Buffer: %p (index=%u)", img_data, this->camera_->get_buffer_index(buffer));
+    ESP_LOGI(TAG, "   Buffer: %p (index=%u)", img_data, (unsigned) this->camera_->get_buffer_index(buffer));
   }
 
   // LVGL 9.4 ZERO-COPY MODE
@@ -289,7 +289,7 @@ void LVGLCameraDisplay::update_canvas_() {
     this->draw_buf_initialized_ = true;
 
     ESP_LOGI(TAG, "Zero-copy draw_buf initialized: %ux%u, stride=%u, size=%u, data=%p",
-             width, height, stride, buf_size, img_data);
+             width, height, (unsigned) stride, (unsigned) buf_size, img_data);
   } else {
     // Just update the data pointer - no memcpy needed!
     this->camera_draw_buf_.data = img_data;
@@ -321,7 +321,7 @@ void LVGLCameraDisplay::configure_canvas(lv_obj_t *canvas) {
   if (canvas != nullptr) {
     lv_coord_t w = lv_obj_get_width(canvas);
     lv_coord_t h = lv_obj_get_height(canvas);
-    ESP_LOGI(TAG, "   Taille canvas: %dx%d", w, h);
+    ESP_LOGI(TAG, "   Taille canvas: %dx%d", (int) w, (int) h);
   }
 }
